# docs/contracts/public-api.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "PersonalTechBrief.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/briefs": {
      "post": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "CreateBrief",
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBriefResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "ListBriefs",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/briefs/current": {
      "get": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "GetCurrentBrief",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/briefs/{id}": {
      "get": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "GetBrief",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/updates/{id}/feedback": {
      "put": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "RecordUpdateFeedback",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/FeedbackRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "ClearUpdateFeedback",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/updates/{id}/saved": {
      "put": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "SaveUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "UnsaveUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/updates/{updateId}/sources/{sourceItemId}/open": {
      "post": {
        "tags": [
          "PersonalTechBrief.Web"
        ],
        "operationId": "RecordSourceOpen",
        "parameters": [
          {
            "name": "updateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sourceItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BriefHistoryResponse": {
        "required": [
          "briefs",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "briefs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BriefSummaryResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BriefItemResponse": {
        "required": [
          "id",
          "technologyUpdateId",
          "rank",
          "title",
          "topic",
          "summary",
          "whyRelevant",
          "relevanceScore",
          "generatedAtUtc",
          "promptVersion",
          "modelOrAlgorithmVersion",
          "sources",
          "currentFeedback",
          "isSaved"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "technologyUpdateId": {
            "type": "string",
            "format": "uuid"
          },
          "rank": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "whyRelevant": {
            "type": "string"
          },
          "relevanceScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "promptVersion": {
            "type": "string"
          },
          "modelOrAlgorithmVersion": {
            "type": "string"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BriefItemSourceResponse"
            }
          },
          "currentFeedback": {
            "type": [
              "null",
              "string"
            ]
          },
          "isSaved": {
            "type": "boolean"
          }
        }
      },
      "BriefItemSourceResponse": {
        "required": [
          "sourceItemId",
          "title",
          "url",
          "publishedAtUtc"
        ],
        "type": "object",
        "properties": {
          "sourceItemId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          },
          "publishedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "BriefResponse": {
        "required": [
          "brief",
          "items"
        ],
        "type": "object",
        "properties": {
          "brief": {
            "$ref": "#/components/schemas/BriefSummaryResponse"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BriefItemResponse"
            }
          }
        }
      },
      "BriefStatus": {
        "enum": [
          "generating",
          "completed",
          "failed"
        ]
      },
      "BriefSummaryResponse": {
        "required": [
          "id",
          "generatedAtUtc",
          "windowStartUtc",
          "windowEndUtc",
          "status",
          "candidateCount",
          "selectedCount",
          "generationVersion",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "windowStartUtc": {
            "type": "string",
            "format": "date-time"
          },
          "windowEndUtc": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/BriefStatus"
          },
          "candidateCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "selectedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "generationVersion": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateBriefResponse": {
        "required": [
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/BriefStatus"
          }
        }
      },
...
      "FeedbackRequest": {
        "required": [
          "value",
          "briefItemId"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "briefItemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}